feat(table): move proptypes to separate folder
