+++ packages/paste-core/components/table/src/Td.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import {Box, safelySpreadBoxProps} from '@twilio-paste/box';
 import type {TdProps} from './types';
-import {TdPropTypes} from './types';
+import {TdPropTypes} from './proptypes';
 
 const Td = React.forwardRef<HTMLTableCellElement, TdProps>(({element = 'TD', textAlign = 'left', ...props}, ref) => {
   return (
