+++ packages/paste-core/components/table/src/Th.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import {Box, safelySpreadBoxProps} from '@twilio-paste/box';
 import type {ThProps} from './types';
-import {ThPropTypes} from './types';
+import {ThPropTypes} from './proptypes';
 
 const Th = React.forwardRef<HTMLTableHeaderCellElement, ThProps>(
   ({element = 'TH', textAlign = 'left', width, ...props}, ref) => {
